+++ app/controllers/concerns/refer/controller.rb
@@ -2,14 +2,22 @@ module Refer
   module Controller
     extend ActiveSupport::Concern
 
-    def set_refer_cookie
-      if (code = params[Refer.param_name])
-        cookies[Refer.cookie_name] = Refer.cookie(code)
+    class_methods do
+      def set_referral_cookie(param_name: Refer.param_name, cookie_name: Refer.cookie_name, **options)
+        before_action -> { set_refer_cookie(param_name: param_name, cookie_name: cookie_name) }, **options
       end
     end
 
-    def refer(referee)
-      Refer.refer(code: cookies[Refer.cookie_name], referee: referee)
+    def refer(referee, cookie_name: Refer.cookie_name)
+      Refer.refer(code: cookies[cookie_name], referee: referee)
+    end
+
+    private
+
+    def set_refer_cookie(param_name: Refer.param_name, cookie_name: Refer.cookie_name)
+      if (code = params[param_name])
+        cookies[cookie_name] = Refer.cookie(code)
+      end
     end
   end
 end