+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### Unreleased
 
+* Add `set_referral_cookie` controller method
+* Add `rails g refer:install` generator to inject `set_referral_cookie`
+
 ### 0.1.1
 
 * Added `dependent: :nullify` so ReferralCodes persist Referral records when deleted.